Make notebooky tests reuse the same cluster - new LeonardoTestUtils trait - Run ClusterMonitorSpec in parallel - Don't need to explicitly mixin WebBrowserSpec - centralize hail file vals - Ensure withOpenNotebook/withNewNotebook always terminate their kernels - Removed redudant test which uploads but doesn't execute notebook
