+++ automation/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/NotebookPage.scala
@@ -15,7 +15,7 @@ class NotebookPage(override val url: String)(override implicit val authToken: Au
   // selects all menus from the header bar
   lazy val menus: Query = cssSelector("[class='dropdown-toggle']")
 
-  // "File" and "Cell" menus
+  // menu elements
 
   lazy val fileMenu: Element = {
     findAll(menus).filter { e => e.text == "File" }.toList.head
@@ -23,6 +23,10 @@ class NotebookPage(override val url: String)(override implicit val authToken: Au
 
   lazy val cellMenu: Element = {
     findAll(menus).filter { e => e.text == "Cell" }.toList.head
+  }
+
+  lazy val kernelMenu: Element = {
+    findAll(menus).filter { e => e.text == "Kernel" }.toList.head
   }
 
   // selects all submenus which appear in dropdowns after clicking a main menu header
@@ -42,6 +46,12 @@ class NotebookPage(override val url: String)(override implicit val authToken: Au
   // Run Cell toolbar button
   lazy val runCellButton: Query = cssSelector("[title='Run']")
 
+  // Kernel -> Shutdown
+  lazy val shutdownKernelSelection: Query = cssSelector("[id='shutdown_kernel']")
+
+  // Kernel -> Shutdown
+  lazy val shutdownKernelConfirmationSelection: Query = cssSelector("[class='btn btn-default btn-sm btn-danger']")
+
   // selects the numbered left-side cell prompts
   lazy val prompts: Query = cssSelector("[class='prompt input_prompt']")
 
@@ -49,6 +59,9 @@ class NotebookPage(override val url: String)(override implicit val authToken: Au
   def cellsAreRunning: Boolean = {
     findAll(prompts).exists { e => e.text == "In [*]:" }
   }
+
+  // can we see that the kernel connection has terminated?
+  lazy val kernelTerminatedModalBody: Query = cssSelector("[class='modal-body']")
 
   def runAllCells(timeoutSeconds: Long): Unit = {
     click on cellMenu
@@ -83,5 +96,12 @@ class NotebookPage(override val url: String)(override implicit val authToken: Au
     click on runCellButton
     await condition (!cellsAreRunning, timeoutSeconds)
     cellOutput(cell)
+  }
+
+  def shutdownKernel(): Unit = {
+    click on kernelMenu
+    click on (await enabled shutdownKernelSelection)
+    click on (await enabled shutdownKernelConfirmationSelection)
+    await enabled kernelTerminatedModalBody
   }
 }
